+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2.4.0 (2018-03-19)
+
+Adds support for new monitor class: `PasteboardMonitor`.
+
+# 2.3.0 (2018-03-15)
+
+Adds support for three new monitor classes: `MetadataQueryMonitor`,
+`UbiquitousKeyValueStoreMonitor`, and `UbiquityIdentityMonitor`.
+
 # 2.2.0 (2018-02-23)
 
 * Adds support for new monitor class: `FileSystemObjectMonitor`.